src/main/java/com/krats/onlinebookstore/repository/BookRepository.java
package com.krats.onlinebookstore.repository;

import com.krats.onlinebookstore.bean.NumberGenerator;
import com.krats.onlinebookstore.bean.TextUtil;
import com.krats.onlinebookstore.model.Book;

import javax.inject.Inject;
import javax.persistence.EntityManager;
import javax.persistence.PersistenceContext;
import javax.persistence.TypedQuery;
import javax.transaction.Transactional;
import javax.validation.constraints.NotNull;
import java.util.List;

@Transactional(Transactional.TxType.SUPPORTS)
public class BookRepository {

    @Inject
    private NumberGenerator numberGenerator;

    @Inject
    private TextUtil textUtil;

    @PersistenceContext(unitName = "onlineBookStorePU")
    private EntityManager em;

    public Book find(@NotNull  Long id) {
        return em.find(Book.class, id);
    }

    @Transactional(Transactional.TxType.REQUIRED)
    public Book create(@NotNull Book book) {
        book.setTitle(textUtil.sanitize(book.getTitle()));
        book.setIsbn(numberGenerator.generateNumber());
        em.persist(book);
        return book;
    }

    @Transactional(Transactional.TxType.REQUIRED)
    public void delete(@NotNull Long id) {
        em.remove(em.getReference(Book.class, id));
    }

    public List<Book> findAll() {
        TypedQuery<Book> query = em.createQuery("SELECT b from Book as b ORDER BY b.title DESC", Book.class);
        return query.getResultList();
    }

    public Long countAll() {
        TypedQuery<Long> query = em.createQuery("SELECT COUNT(b) FROM Book b", Long.class);
        return query.getSingleResult();
    }

}
